+++ mcp-server/src/toolProviders/debug/index.ts
@@ -1,8 +1,9 @@
 export * from './getConfigurations';
 export * from './setBreakpoint';
 export * from './getBreakpoints';
+export * from './getDebugState';
 export * from './removeBreakpoint';
 export * from './startDebugging';
 export * from './continueDebugging';
-export { stepExecutionTool } from './stepExecution';
+export * from './stepExecution';
 export * from './stopDebugging';